Change int data type to float data type

+++ src/MyPackage/Add.java
@@ -1,7 +1,7 @@
 package MyPackage;
 
 public class Add {
-  protected static void AddTwoNumbers(int num1, int num2) {
+  protected static void AddTwoNumbers(float num1, float num2) {
     System.out.println(num1 + num2);
   }
 }
